web/data_manager/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
pyyaml>=6.0
watchdog>=4.0
aiofiles>=23.2
numpy>=1.26  # required by rclpy sensor_msgs
Pillow>=10.0  # JPEG encoding for camera MJPEG endpoint
av>=12.0  # PyAV: AV1/H.264 mp4 encoding for episode recording
pyarrow>=15.0  # LeRobot v2.1 Parquet writer
zarr>=2.18,<3  # Depth (uint16) per-episode chunked store, training-friendly
